Add new PreView Image

+++ CarSharingApplication/DataTransferObjects/CreateCarSharingProfileModelObject.cs
@@ -15,7 +15,7 @@ namespace CarSharingApplication.DataTransferObjects
         //Car Profile
         public string Name { get; set; } = default!;
         public string? Description { get; set; }
-        public decimal? PricePerDay { get; set; }
+        public int? PricePerDay { get; set; }
         public List<IFormFile> Images { get; set; } = default!;
 
         //Contact Details
